urandom: don't seed/store in containers

+++ core-services/05-misc.sh
@@ -1,8 +1,10 @@
 # vim: set ts=4 sw=4 et:
 
-msg "Initializing random seed..."
-cp /var/lib/random-seed /dev/urandom >/dev/null 2>&1 || true
-( umask 077; bytes=$(cat /proc/sys/kernel/random/poolsize) || bytes=512; dd if=/dev/urandom of=/var/lib/random-seed count=1 bs=$bytes >/dev/null 2>&1 )
+if [ -z "$VIRTUALIZATION" ]; then
+    msg "Initializing random seed..."
+    cp /var/lib/random-seed /dev/urandom >/dev/null 2>&1 || true
+    ( umask 077; bytes=$(cat /proc/sys/kernel/random/poolsize) || bytes=512; dd if=/dev/urandom of=/var/lib/random-seed count=1 bs=$bytes >/dev/null 2>&1 )
+fi
 
 msg "Setting up loopback interface..."
 ip link set up dev lo
